Remove unnecessary casts

+++ src/test/java/de/nihas101/chip8/unsignedDataTypes/UnsignedDataTypeTest.java
@@ -8,32 +8,32 @@ public class UnsignedDataTypeTest {
 
     @Test
     public void equals() {
-        UnsignedByte unsignedByte1 = new UnsignedByte((byte) 10);
-        UnsignedByte unsignedByte2 = new UnsignedByte((byte) 10);
+        UnsignedByte unsignedByte1 = new UnsignedByte(10);
+        UnsignedByte unsignedByte2 = new UnsignedByte(10);
 
         assertEquals(true, unsignedByte1.equals(unsignedByte2));
     }
 
     @Test
     public void compareTo() {
-        UnsignedByte unsignedByte1 = new UnsignedByte((byte) 10);
-        UnsignedByte unsignedByte2 = new UnsignedByte((byte) 15);
+        UnsignedByte unsignedByte1 = new UnsignedByte(10);
+        UnsignedByte unsignedByte2 = new UnsignedByte(15);
 
         assertEquals(5, unsignedByte2.compareTo(unsignedByte1));
     }
 
     @Test
     public void apply() {
-        UnsignedByte unsignedByte = new UnsignedByte((byte) 10);
+        UnsignedByte unsignedByte = new UnsignedByte(10);
 
-        assertEquals(5, unsignedByte.apply((x, y) -> x - y, new UnsignedByte((byte) 5)).unsignedDataType);
+        assertEquals(5, unsignedByte.apply((x, y) -> x - y, new UnsignedByte(5)).unsignedDataType);
     }
 
     @Test
     public void apply1() {
-        UnsignedShort unsignedShort = new UnsignedShort((byte) 10);
+        UnsignedShort unsignedShort = new UnsignedShort(10);
 
-        assertEquals(5, unsignedShort.apply((x, y) -> x - y, new UnsignedShort((byte) 5)).unsignedDataType);
+        assertEquals(5, unsignedShort.apply((x, y) -> x - y, new UnsignedShort(5)).unsignedDataType);
     }
 
     @Test
